+++ contracts/fx-portal/tunnel/ERC20/FxERC20ChildTunnel.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity ^0.8.8;
+pragma solidity ^0.8.28;
 
 import {IFxERC20} from "./../../token/ERC20/interfaces/IFxERC20.sol";
 import {IForwarderRegistry} from "@animoca/ethereum-contracts/contracts/metatx/interfaces/IForwarderRegistry.sol";