+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Upgraded to `@openzeppelin/contracts@5.2.0`.
 - Upgraded to `solc@0.8.28`.
 - Upgraded to `@animoca/ethereum-contracts@4.0.0`.
-- change variable casing for public immutable `FxERC20ChildTunnel`.`CHILD_TOKEN_LOGIC` and `FxERC20RootTunnel`.`CHILD_TOKEN_PROXY_CODE_HASH`.
+- Changed public variables casing for `FxERC20ChildTunnel`.`CHILD_TOKEN_LOGIC` and `FxERC20RootTunnel`.`CHILD_TOKEN_PROXY_CODE_HASH`.
 
 ### New features